Se agrega funcionalidad sin terminar del Home
