+++ app/src/main/java/com/example/cookbook/MainActivity.kt
@@ -10,9 +10,7 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.navigation.compose.rememberNavController
 import com.example.cookbook.navigation.MyAppNavigationView
-import com.example.cookbook.presentation.home.view.HomeView
 import com.example.cookbook.ui.theme.CookBookTheme
 import com.google.firebase.Firebase
 import com.google.firebase.messaging.messaging
