+++ app/src/main/java/com/example/cookbook/presentation/finder/network/IngredientByCategory.kt
@@ -1,0 +1,5 @@
+package com.example.cookbook.presentation.finder.network
+
+object IngredientByCategory {
+
+}
