+++ app/src/main/java/com/example/cookbook/presentation/home/home/viewmodels/HomeViewModelFactory.kt
@@ -1,0 +1,15 @@
+package com.example.cookbook.presentation.home.home.viewmodels
+
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
+import com.example.cookbook.presentation.home.home.network.HomeRepository
+
+class HomeViewModelFactory (private val repository: HomeRepository) : ViewModelProvider.Factory{
+    override fun <T : ViewModel> create(modelClass: Class<T>): T {
+        if (modelClass.isAssignableFrom(HomeViewModel::class.java)) {
+            @Suppress("UNCHECKED_CAST")
+            return HomeViewModel(repository) as T
+        }
+        throw IllegalArgumentException("Unknown ViewModel class")
+    }
+}
