--- app/src/main/java/com/example/cookbook/presentation/home/view/HomeViewModel.kt
@@ -1,4 +1,0 @@
-package com.example.cookbook.presentation.home.view
-
-class HomeViewModel {
-}
